parse nums from hex fun FormatTransaction

+++ src/blockchainHandlers/struct.go
@@ -7,11 +7,11 @@ type Transaction struct {
 	Nonce            int    `json:"nonce"`
 	To               string `json:"to"`
 	TransactionIndex int    `json:"transactionIndex"`
-	Value            int64  `json:"value"`
+	Value            int  `json:"value"`
 	GasUsed          int    `json:"gas_used"`
 	BlockHash        string `json:"block_hash"`
 	BlockNumber      int    `json:"block_number"`
-	GasPrice         int64  `json:"gas_price"`
+	GasPrice         int  `json:"gas_price"`
 	Timestamp        int    `json:"timestamp"`
 	BlockMiner       string `json:"block_miner"`
 	Journal          [] struct {
